# docs/openapi.json
{
 "openapi": "3.0.1",
 "info": {
  "title": "OpenAPI definition",
  "version": "v0"
 },
 "servers": [
  {
   "url": "http://localhost:18090",
   "description": "Generated server url"
  }
 ],
 "paths": {
...
  "/admin/services": {
   "get": {
    "tags": [
     "admin-controller"
    ],
    "operationId": "services_2",
    "parameters": [
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "format": "int32"
      }
     },
     {
      "name": "size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "format": "int32"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "array",
         "items": {
          "type": "object",
          "additionalProperties": {
           "type": "object"
          }
         }
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "admin-controller"
    ],
    "operationId": "createService",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ServiceReq"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "object",
         "additionalProperties": {
          "type": "object"
         }
        }
       }
      }
     }
    }
   }
  },
...
  "/admin/partners": {
   "get": {
    "tags": [
     "admin-controller"
    ],
    "operationId": "partners",
    "parameters": [
     {
      "name": "page",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "format": "int32"
      }
     },
     {
      "name": "size",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "format": "int32"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "array",
         "items": {
          "type": "object",
          "additionalProperties": {
           "type": "object"
          }
         }
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "admin-controller"
    ],
    "operationId": "createPartner",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PartnerReq"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "object",
         "additionalProperties": {
          "type": "object"
         }
        }
       }
      }
     }
    }
   }
  },
...
  "/admin/billing/periods": {
   "get": {
    "tags": [
     "billing-query-controller"
    ],
    "operationId": "periods",
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "array",
         "items": {
          "type": "object",
          "additionalProperties": {
           "type": "object"
          }
         }
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "billing-controller"
    ],
    "operationId": "close_1",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PeriodReq"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "object",
         "additionalProperties": {
          "type": "object"
         }
        }
       }
      }
     }
    }
   }
  },
  "/admin/billing/payouts": {
   "get": {
    "tags": [
     "billing-query-controller"
    ],
    "operationId": "payouts_1",
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "array",
         "items": {
          "type": "object",
          "additionalProperties": {
           "type": "object"
          }
         }
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "billing-controller"
    ],
    "operationId": "createPayout",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PayoutReq"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "object",
         "additionalProperties": {
          "type": "object"
         }
        }
       }
      }
     }
    }
   }
  },
...
  "/portal/statements": {
   "get": {
    "tags": [
     "portal-query-controller"
    ],
    "operationId": "statements",
    "parameters": [
     {
      "name": "from",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "format": "date-time"
      }
     },
     {
      "name": "to",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "format": "date-time"
      }
     },
     {
      "name": "format",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "json"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "object"
        }
       }
      }
     }
    }
   }
  },
...
  "/portal/results": {
   "get": {
    "tags": [
     "portal-query-controller"
    ],
    "operationId": "results",
    "parameters": [
     {
      "name": "serviceId",
      "in": "query",
      "required": true,
      "schema": {
       "type": "integer",
       "format": "int64"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "array",
         "items": {
          "type": "object",
          "additionalProperties": {
           "type": "object"
          }
         }
        }
       }
      }
     }
    }
   }
  },
...
  "/portal/export": {
   "get": {
    "tags": [
     "portal-query-controller"
    ],
    "operationId": "export",
    "parameters": [
     {
      "name": "format",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "csv"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "string",
         "format": "byte"
        }
       }
      }
     }
    }
   }
  },
...
  "/admin/services/{id}/campaign": {
   "get": {
    "tags": [
     "campaign-query-controller"
    ],
    "operationId": "campaign",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "format": "int64"
      }
     },
     {
      "name": "hours",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "format": "int32",
       "default": 24
      }
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "object",
         "additionalProperties": {
          "type": "object"
         }
        }
       }
      }
     }
    }
   }
  },
  "/admin/services/{id}/campaign/export": {
   "get": {
    "tags": [
     "campaign-query-controller"
    ],
    "operationId": "campaignExport",
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "format": "int64"
      }
     },
     {
      "name": "format",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "pdf"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "string",
         "format": "byte"
        }
       }
      }
     }
    }
   }
  },
...
  "/admin/billing/statements": {
   "get": {
    "tags": [
     "billing-query-controller"
    ],
    "operationId": "statement",
    "parameters": [
     {
      "name": "partnerId",
      "in": "query",
      "required": true,
      "schema": {
       "type": "integer",
       "format": "int64"
      }
     },
     {
      "name": "from",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "format": "date-time"
      }
     },
     {
      "name": "to",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "format": "date-time"
      }
     },
     {
      "name": "format",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "json"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "OK",
      "content": {
       "*/*": {
        "schema": {
         "type": "object"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "ServiceReq": {
    "type": "object",
    "properties": {
     "name": {
      "type": "string"
     },
     "type": {
      "type": "string",
      "enum": [
       "VOTE",
       "QUIZ",
       "PREMIUM_CONTENT",
       "SUBSCRIPTION",
       "ALERT"
      ]
     },
     "partnerId": {
      "type": "integer",
      "format": "int64"
     },
     "shortCodeId": {
      "type": "integer",
      "format": "int64"
     },
     "regulated": {
      "type": "boolean"
     },
     "consentMode": {
      "type": "string",
      "enum": [
       "SIMPLE_OPT_IN",
       "DOUBLE_OPT_IN",
       "API_ACTIVATION"
      ]
     },
     "opensAt": {
      "type": "string",
      "format": "date-time"
     },
     "closesAt": {
      "type": "string",
      "format": "date-time"
     },
     "maxActionsPerMsisdn": {
      "type": "integer",
      "format": "int32"
     },
     "defaultLang": {
      "type": "string"
     },
     "replyOk": {
      "type": "string"
     },
     "replyStop": {
      "type": "string"
     },
     "replyHelp": {
      "type": "string"
     },
     "replyLimit": {
      "type": "string"
     },
     "replyClosed": {
      "type": "string"
     },
     "maxTps": {
      "type": "integer",
      "format": "int32"
     }
    }
   },
...
   "PartnerReq": {
    "type": "object",
    "properties": {
     "name": {
      "type": "string"
     },
     "sharePercent": {
      "type": "number"
     },
     "webhookUrl": {
      "type": "string"
     },
     "webhookSecret": {
      "type": "string"
     },
     "maxTps": {
      "type": "integer",
      "format": "int32"
     }
    }
   },
...
   "PeriodReq": {
    "type": "object",
    "properties": {
     "from": {
      "type": "string",
      "format": "date-time"
     },
     "to": {
      "type": "string",
      "format": "date-time"
     }
    }
   },
   "PayoutReq": {
    "type": "object",
    "properties": {
     "partnerId": {
      "type": "integer",
      "format": "int64"
     },
     "from": {
      "type": "string",
      "format": "date-time"
     },
     "to": {
      "type": "string",
      "format": "date-time"
     }
    }
   },
...
  }
 }
}